Move customized JSON encoder to separate module. Use encoder to string representation of Error object
