+++ aiohttp_json_api/utils.py
@@ -13,20 +13,12 @@ from aiohttp import web
 from boltons.iterutils import remap, first
 from boltons.typeutils import make_sentinel
 
-from .jsonpointer import JSONPointer
-from .helpers import is_collection
 from .const import JSONAPI, JSONAPI_CONTENT_TYPE
+from .helpers import is_collection
+from .encoder import json_dumps
 from .errors import Error, ErrorList, ValidationError
 
 SENTINEL = make_sentinel()
-
-
-class JSONEncoder(json.JSONEncoder):
-    def default(self, o):
-        if isinstance(o, JSONPointer):
-            return o.path
-        else:
-            return super(JSONEncoder, self).default(o)
 
 
 def filter_empty_fields(data: typing.MutableMapping) -> typing.MutableMapping:
@@ -43,7 +35,7 @@ def jsonapi_response(data=SENTINEL, *, text=None, body=None,
                      reason=None, headers=None,
                      dumps=None):
     if not callable(dumps):
-        dumps = functools.partial(json.dumps, cls=JSONEncoder)
+        dumps = json_dumps
 
     return web.json_response(
         data=data, text=text, body=body, status=status,
